[Refactoring] extract immediate worker from RabbitMQWorkqueueTest
